+++ CHANGELOG.md
@@ -4,10 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[1.1.0] - 2024-06-07
+[1.2.0] - 2024-06-17
 
 ### Fixed
-Send event properties as third param in trackEvent() call
+ - Fixed segmentation evaluator issues where `contains` and `Greater than equal to` operator were not working as expected.
+ - Added unit test cases for the segmentation evaluator and decision maker
+
+`[1.1.0] - 2024-06-07
+
+### Fixed
+Send event properties as third param in trackEvent() call`
 
 [1.0.0] - 2024-05-31